update the files structures
