+++ src/component/App.js
@@ -1,9 +1,13 @@
 import React from 'react';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
-import Home from './Home';
-import LogIn from './Login';
-import Register from './Signup';
-import Welcome from './Welcome';
+import Home from './Home/Home';
+import LogIn from './Home/Login';
+import Register from './Home/Signup';
+import Welcome from './Main/Welcome';
+import Reserve from './Main/Forms/Reserve';
+import MyReservation from './Main/Forms/MyReservation';
+import AddCar from './Main/Forms/AddCar';
+import RemoveCar from './Main/Forms/RemoveCar';
 
 function App() {
   return (
@@ -13,11 +17,14 @@ function App() {
           <Route path="/" element={<Home />} />
           <Route path="/login" element={<LogIn />} />
           <Route path="/signup" element={<Register />} />
-          <Route path="/welcome" element={<Welcome />} />
+          <Route path="/main" element={<Welcome />} />
+          <Route path="/reserve" element={<Reserve />} />
+          <Route path="/my-reservations" element={<MyReservation />} />
+          <Route path="/add-a-car" element={<AddCar />} />
+          <Route path="/remove-a-car" element={<RemoveCar />} />
         </Routes>
       </BrowserRouter>
     </div>
   );
 }
-
 export default App;
